got debug_env running in tune

+++ env.py
@@ -205,10 +205,12 @@ class Env(gym.Env):
                         * self.n_lines
                     )
                 ),
-                obs=spaces.Box(low=0, high=1, shape=self.world_shape),
+                obs=spaces.Box(low=0, high=1, shape=self.world_shape, dtype=np.float32),
             )._asdict()
         )
-        self.world_space = spaces.Box(low=0, high=self.world_size - 1, shape=[2])
+        self.world_space = spaces.Box(
+            low=0, high=self.world_size - 1, shape=[2], dtype=np.float32
+        )
 
     @staticmethod
     @functools.lru_cache(maxsize=200)
@@ -603,7 +605,12 @@ class Env(gym.Env):
         return self.iterator.send(action)
 
     def render_world(
-        self, state, action, lower_level_action, reward, cumulative_reward,
+        self,
+        state,
+        action,
+        lower_level_action,
+        reward,
+        cumulative_reward,
     ):
 
         if action is not None and action < len(self.subtasks):
@@ -611,7 +618,8 @@ class Env(gym.Env):
         print("Action:", action)
         if lower_level_action is not None:
             print(
-                "Lower Level Action:", self.lower_level_actions[lower_level_action],
+                "Lower Level Action:",
+                self.lower_level_actions[lower_level_action],
             )
         print("Reward", reward)
         print("Cumulative", cumulative_reward)
@@ -639,7 +647,12 @@ class Env(gym.Env):
             print("-" * len(string))
 
     def render_instruction(
-        self, term, success, lines, state, agent_ptr,
+        self,
+        term,
+        success,
+        lines,
+        state,
+        agent_ptr,
     ):
 
         if term:
